+++ src/main/java/com/ProjectoJava/objetos/entity/OrderLine.java
@@ -1,13 +1,13 @@
 package com.ProjectoJava.objetos.entity;
 
 public class OrderLine {
-    int idProducto;
+    Long idProducto;
     int cantidad;
-    public OrderLine(int idProducto, int unaCantidad){
+    public OrderLine(Long idProducto, int unaCantidad){
         this.idProducto = idProducto;
         this.cantidad = unaCantidad;
     }
 
-    public int getIdProducto(){ return this.idProducto;}
+    public Long getIdProducto(){ return this.idProducto;}
     public int getCantidad(){return this.cantidad;}
 }
